fix(cockroachdb): retain existing checklist index

+++ api/migrations/schema/Version20240928105306.php
@@ -18,6 +18,10 @@ final class Version20240928105306 extends AbstractMigration {
     }
 
     public function up(Schema $schema): void {
+        if (CockroachDb::is($this->connection)) {
+            return;
+        }
+
         $this->addSql(
             <<<'EOF'
                 drop index checklistitem_checklistid_parentid_position_unique
@@ -36,6 +40,10 @@ final class Version20240928105306 extends AbstractMigration {
 
     #[\Override]
     public function down(Schema $schema): void {
+        if (CockroachDb::is($this->connection)) {
+            return;
+        }
+
         $this->addSql(
             <<<'EOF'
                 alter table checklist_item
